library-management-system/backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.35
psycopg==3.2.3
alembic==1.13.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration & Environment
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Web Scraping
beautifulsoup4==4.12.2
lxml==4.9.3

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Logging
structlog==23.2.0

# Monitoring & Metrics
prometheus-client==0.22.0
psutil==7.0.0

# CORS (FastAPIに組み込まれているCORSミドルウェアを使用)
